Increase declared Perl dependency to 5.8
Was previously 5.6, but it's impossible to run the tests on a clean 5.6
install: the tests use Test::Most, which depends on Exception::Class, which
won't install out-of-the-box on 5.6.2.

Since 5.6 actually isn't an option (and, I assume, this dist hasn't actually
been tested on 5.6), it seems better to make that decision now, and permit
use of 5.8-specific features (even if the code doesn't currently use any).